LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-first-002-expected.txt




























FAIL .target > * 1 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 30
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 10
FAIL .target > * 4 assert_equals:
<div class="inner" style="overflow: hidden; display: inline-block;" data-offset-x="10">
    <div><span></span></div>
    <div><span></span></div>
  </div>
offsetLeft expected 10 but got 20
FAIL .target > * 5 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 45
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<span data-offset-x="65"></span>
offsetLeft expected 65 but got 10
FAIL .target > * 8 assert_equals:
<div class="inner" style="display: inline-block; columns: 2;" data-offset-x="10">
    <div style="font-size: 20px;"><span></span></div>
    <div style="font-size: 20px;"><span></span></div>
    <div><span></span></div>
    <div><span></span></div>
  </div>
offsetLeft expected 10 but got 53
FAIL .target > * 9 assert_equals:
<span data-offset-x="90"></span>
offsetLeft expected 90 but got 30
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 45
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 45
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 100
PASS .target > * 16
FAIL .target > * 17 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 57
PASS .target > * 18
PASS .target > * 19
PASS .target > * 20
FAIL .target > * 21 assert_equals:
<span data-offset-x="160"></span>
offsetLeft expected 160 but got 10
FAIL .target > * 22 assert_equals:
<div class="inner" style="display: inline-block;" data-offset-x="10">
    <table>
      <caption></caption>
      <tbody><tr>
        <td style="font-size: 10px;"><span></span><br><span></span></td>
        <td><span></span><br><span></span></td>
      </tr>
      <tr>
        <td><span></span><br><span></span></td>
        <td style="font-size: 10px;"><span></span><br><span></span></td>
      </tr>
      </tbody><caption style="caption-side: bottom;"></caption>
    </table>
  </div>
offsetLeft expected 10 but got 20
